Move pipeline under text directly.
